pyproject.toml
[project]
name = "lactationcurve"
version = "0.1.0"
description = "Lactation curve fitting and analysis library"
readme = "README.md"
requires-python = ">= 3.11, < 3.12"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "scipy>=1.16.3",
    "sympy>=1.14.0",
    "matplotlib>=3.10.8",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lactationcurve"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
